backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
fastapi==0.109.1
groq==0.4.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.1
starlette==0.35.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.0
